adionada menu na pagina de Vacinas pelo mundo e corrigido erros do backend

+++ dose-em-dia-backend/DoseEmDia/DoseEmDia/Models/Vacina.cs
@@ -12,7 +12,7 @@ namespace DoseEmDia.Models
         public string Nome { get; set; }
         public string Fabricante { get; set; }
         public int NumeroDoses { get; set; }
-        public long NumeroLote { get; set; }
+        public int NumeroLote { get; set; }
         public string IntervaloEntreDoses { get; set; }
         public DateTime DataAplicacao { get; set; }
         public StatusVacina Status { get; set; }
@@ -20,7 +20,7 @@ namespace DoseEmDia.Models
         public int UsuarioId { get; set; }
         public int? ValidadeMeses { get; set; }
 
-        public Vacina(string nome, string fabricante,int numeroDoses, long numeroLote, string intervaloEntreDoses, DateTime dataAplicacao, StatusVacina status)
+        public Vacina(string nome, string fabricante,int numeroDoses, int numeroLote, string intervaloEntreDoses, DateTime dataAplicacao, StatusVacina status)
         {
             Nome = nome;
             Fabricante = fabricante;
